requirements.txt
aiohttp==3.5.1
async-timeout==3.0.1
attrs==18.2.0
backcall==0.1.0
boto==2.49.0
boto3==1.9.80
botocore==1.12.80
bz2file==0.98
certifi==2018.11.29
chardet==3.0.4
decorator==4.3.0
DingtalkChatbot==1.3.0
docutils==0.14
gensim==3.6.0
grpcio==1.16.0
grpcio-tools==1.16.0
idna==2.7
idna-ssl==1.1.0
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.2
jieba==0.39
jmespath==0.9.3
JPype1==0.6.3
kafka==1.3.5
kazoo==2.5.0
multidict==4.5.2
nltk==3.4
numpy==1.14.3
pandas==0.22.0
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
pika==0.13.0
pip==18.1
prompt-toolkit==2.0.7
protobuf==3.6.0
ptyprocess==0.6.0
py4j==0.10.7
Pygments==2.3.1
pykafka==2.8.0
PyMySQL==0.7.9
pyspark==2.4.0
python-crfsuite==0.9.6
python-dateutil==2.7.5
pytz==2018.9
PyYAML==3.13
redis==2.10.6
redis-py-cluster==1.3.6
requests==2.20.1
s3transfer==0.1.13
scikit-learn==0.20.2
scipy==1.2.0
setuptools==40.6.3
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
sklearn-crfsuite==0.3.6
smart-open==1.8.0
tabulate==0.8.2
tornado==5.1.1
tqdm==4.29.0
traitlets==4.3.2
typing-extensions==3.6.6
ujson==1.35
urllib3==1.24.1
wasp-eureka==1.1.1
wcwidth==0.1.7
wheel==0.32.3
yarl==1.3.0
